Add reviews/isReviewedBy to deposit form and render
and example of a review with DOI from PeerJ
https://doi.org/10.7287/peerj.5551v0.2/reviews/1

Adding this will allow us to update a record to link to DOI and HTTPs of
what has reviewed say a preprint;

or to deposit a review and indicate what paper/preprint it is reviewing

Proper Open Science! :)

See DataCite relations:
https://github.com/zenodo/zenodo/blob/master/zenodo/modules/records/data/include/v4.1/datacite-relationType-v4.1.xsd#L34

(perhaps others of these should also be considered)

WIP - how do I ensure the relation makes it way into DataCite
registration?

Should look similar to (is-review-of) in
http://api.crossref.org/v1/works/10.7287/peerj.5551v0.2/reviews/1